+++ posthog/models/person/person.py
@@ -572,6 +572,9 @@ def get_distinct_ids_for_subquery(person: Person | None, team: Team) -> list[str
     last_ids_limit = MAX_LIMIT_DISTINCT_IDS - first_ids_limit
 
     if person is not None:
+        # When a Person comes from personhog (via proto_person_to_model), distinct IDs
+        # are already populated on _distinct_ids — use them directly to avoid hitting
+        # the Django ORM below, which would defeat the purpose of the personhog path.
         if hasattr(person, "_distinct_ids") and person._distinct_ids is not None:
             ids = person._distinct_ids
             if len(ids) <= MAX_LIMIT_DISTINCT_IDS:
